+++ src/pf1/sparse_index.rs
@@ -62,7 +62,7 @@ impl DeserializeFromCpp for SparseIndex {
             let accum_lens = EFVector::from_usize_slice(&accum_lens)?;
 
             let unitigs = UnitigSet {
-                k: base.k as usize,
+                k: base.k,
                 useq,
                 accum_lens,
                 bv,
